// cottage-app/src/pages/portfolio/types/Project.ts
interface Project {
    demoImage: string; // e.g. https://storage.googleapis.com/cottage-assets/cragslist.PNG
    name: string; // e.g. Craigslist Redesign
    description: string; // e.g. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididun.
    link: string;
    recordId: string;
  }
  
  export default Project;
  